requirements.txt
# License Key

py-machineid

# HTTP Request

httpx
